mouse moves on desktop [Buggy]
The mouse visually stutters. This should be because we are occasionally displaying frames before they are fully rendered.

It would be best to have something like a desktop.render() method that creates a desired frame, before passing that frame to a desktop.Draw() that actually displays that frame or something similar.